refactor:adjust the mapstr code structure, issue #1625
