+++ src/common/mapstr/mapstr.go
@@ -36,16 +36,15 @@ func (cli MapStr) ToMapInterface() map[string]interface{} {
 	return cli
 }
 
-/*ToStructByTag convert self into a struct with 'tagName'
-
-  eg:
-  self := MapStr{"testName":"testvalue"}
-  targetStruct := struct{
-      Name string `field:"testName"`
-  }
-  After call the function self.ToStructByTag(targetStruct, "field")
-  the targetStruct.Name value will be 'testvalue'
-*/
+// ToStructByTag convert self into a struct with 'tagName'
+//
+//  eg:
+//  self := MapStr{"testName":"testvalue"}
+//  targetStruct := struct{
+//      Name string `field:"testName"`
+//  }
+//  After call the function self.ToStructByTag(targetStruct, "field")
+//  the targetStruct.Name value will be 'testvalue'
 func (cli MapStr) ToStructByTag(targetStruct interface{}, tagName string) error {
 	return SetValueToStructByTagsWithTagName(targetStruct, cli, tagName)
 }
